backend/requirements.txt
anyascii==0.3.0
asgiref==3.5.0
autopep8==1.6.0
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-cors-headers==3.11.0
django-filter==21.1
django-modelcluster==5.3
django-taggit==2.1.0
django-treebeard==4.5.1
djangorestframework==3.13.1
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
Jinja2==3.0.3
l18n==2021.3
Markdown==3.3.6
MarkupSafe==2.1.0
openpyxl==3.0.9
optional-django==0.3.0
Pillow==9.0.1
pycodestyle==2.8.0
pyembed==1.3.3
pyembed-markdown==1.1.0
Pygments==2.11.2
pymdown-extensions==9.2
pytz==2021.3
PyYAML==6.0
react==4.3.0
requests==2.27.1
scripts==2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stua==0.2
tablib==3.2.0
telepath==0.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.8
wagtail==2.16.1
webencodings==0.5.1
Willow==1.4.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zipp==3.7.0
gunicorn==20.1.0
mysqlclient==2.1.0
